+++ CHANGELOG.md
@@ -23,3 +23,4 @@ It's the first version.
 * Fixed documentation link
 * Bugfixes relating to inheritance
 * Auto-enable builtin tokens
+* Fix beat signal staying on when paused