+++ CHANGELOG.md
@@ -22,3 +22,5 @@ It's the first version.
     * Key
 * Fixed documentation link
 * Bugfixes relating to inheritance
+* Auto-enable builtin tokens
+* Fix beat signal staying on when paused